Many updates to the ReactCropperJS.jsx component
Many updates have been made to the ReactCropperJS.jsx component. Most
notably the callback methods like crop and cropend now pass the event
and the data URL of the image to the callback function passed in to the
component. This also updates the props list in the ReactCropperJS.jsx
component so that only options that the CropperJS library supports can
be passed in.