Update Readme, Remove Unneeded Files
Updates the README with information on props and usage. Also, removes
dist folder since anyone using the component will be building their
apps, and it seems to be the standard in react component repositories.
Styles were moved to the example folder since that's the only place they
will be used. The user of the component will need to include the css
file in their app from the node_modules/cropperjs/dist folder. Removes
the dependencies for url-loader and file-loader since we aren't loading
any images for the component.